third_party/WebKit/LayoutTests/media/track/track-webvtt-tc003-newlines.html

<!DOCTYPE html>
<title>Tests that a cue with no newline at eof is properly parsed.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track src="captions-webvtt/tc003-no-newline-at-eof.vtt" default>
</video>
<script>
async_test(function(t) {
    var track = document.querySelector("track");

    track.onload = t.step_func_done(function() {
        var expected = [
            {
                id : "1",
                startTime : 0,
                endTime : 30.5,
                text : "Bear is Coming!!!!!"
            }
        ];

        assert_cues_equal(track.track.cues, expected);
    });
});
</script>
